+++ src/Module.php
@@ -13,6 +13,10 @@ class Module extends \luya\base\Module
     const VERSION = '1.0.0-alpha6';
     
     /**
+     * @todo rename to: compositionPattern, compisitionHidden, compositionDefault
+     */
+    
+    /**
      * Url matching prefix, which is used for all the modules (e.g. an e-store requireds a language
      * as the cms needs this informations too). After proccessing this informations, they will be removed
      * from the url for further proccessing.