+++ CHANGELOG.md
@@ -4,6 +4,7 @@ LUYA CHANGELOG
 1.0.0-alpha6 (in progress)
 --------------------------
 - removed `$app->collection->links` (search for `$app->collection->links` and replace with `$app->links`).
+- removed `$app->collection->composition` (search for `$app->collection->composition` and replace with `$app->collection`).
 - added luya version constante to module `luya\Module::VERSION`.
 - added user online overview in admin panel.
 - fixed `#64` wrong logout url.